refactor(harness): split pool/partition operator leaf from transport trunk
Demote partition/smoke/claim to batch_operator.py; batch_differential keeps the
ship-and-run transport. Re-exported so existing imports/tests are unchanged.
